Split screens package into screens for each use case
