+++ src/main/java/event_creation_screens/EventCreationController.java
@@ -1,4 +1,4 @@
-package screens;
+package event_creation_screens;
 
 import event_creation_use_case.EventCreationInputBoundary;
 import event_creation_use_case.EventCreationRequestModel;
