+++ src/main/java/event_creation_screens/EventCreationFailed.java
@@ -1,4 +1,4 @@
-package screens;
+package event_creation_screens;
 
 public class EventCreationFailed extends RuntimeException {
     public EventCreationFailed(String error) {
